* Resolved conflicts in makefile between makefile improvements (from
  master) and additions supporting main (from dev/afbarnard).

+++ java/src/mloss/roc/Curve.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2013 Roc Project.  This is free software.  See
+ * Copyright (c) 2014 Roc Project.  This is free software.  See
  * LICENSE.txt for details.
  */
 
@@ -318,7 +318,16 @@ public class Curve {
          * corresponds to the area under the curve.  (I'm not sure why.)
          */
         double[] uStatistics = mannWhitneyU();
-        return uStatistics[1] / (double) (totalPositives * totalNegatives);
+        // Cast totalPositives and totalNegatives to double to avoid
+        // integer overflow in the product.  For our supported scale
+        // of up to 2^20 (~ 1 million) examples the multiplication
+        // will not lose precision because the factors are at most
+        // 2^20 and the product will be at most 2^40 which fits in a
+        // double's 53 bits of precision. For larger data sizes there
+        // should be minimal floating point error provided
+        // totalPositives and totalNegatives are not on extremely
+        // different scales (i.e. the exponents are similar).
+        return uStatistics[1] / ((double) totalPositives * (double) totalNegatives);
     }
 
     /**
@@ -689,8 +698,17 @@ public class Curve {
             sumPosRanks += rank * (double) posCount;
             sumNegRanks += rank * (double) negCount;
         }
-        double uPos = sumPosRanks - (double) (totalPositives * totalPositives + totalPositives) / 2.0;
-        double uNeg = sumNegRanks - (double) (totalNegatives * totalNegatives + totalNegatives) / 2.0;
+        // Cast totalPositives and totalNegatives to double to avoid
+        // integer overflow in the product.  For our supported scale
+        // of up to 2^20 (~ 1 million) examples the multiplication
+        // will not lose precision because the factors are at most
+        // 2^20 and the product will be at most 2^40 which fits in a
+        // double's 53 bits of precision. For larger data sizes there
+        // should be minimal floating point error provided
+        // totalPositives and totalNegatives are not on extremely
+        // different scales (i.e. the exponents are similar).
+        double uPos = sumPosRanks - ((double) totalPositives * (double) totalPositives + (double) totalPositives) / 2.0;
+        double uNeg = sumNegRanks - ((double) totalNegatives * (double) totalNegatives + (double) totalNegatives) / 2.0;
         return new double[] {uPos, uNeg};
     }
